refactor: domain 하위 패키지 추가
