+++ src/main/java/domain/blackjack/AbstractRandomCardDrawStrategy.java
@@ -1,5 +1,7 @@
-package domain;
+package domain.blackjack;
 
+import domain.card.Card;
+import domain.card.CardDrawStrategy;
 import java.util.List;
 import java.util.Random;
 
@@ -12,7 +14,7 @@ public abstract class AbstractRandomCardDrawStrategy implements CardDrawStrategy
         throw new IllegalStateException("카드를 더이상 뽑을 수 없습니다.");
     }
 
-    protected abstract boolean canDraw();
+    public abstract boolean canDraw();
 
     private Card cardSelectStrategy(List<Card> cards) {
         Random random = new Random();
