+++ src/main/java/domain/blackjack/CardPoint.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.blackjack;
 
 public record CardPoint(int point) {
