+++ src/main/java/domain/blackjack/CardPointCalculator.java
@@ -1,4 +1,8 @@
-package domain;
+package domain.blackjack;
+
+import domain.blackjack.CardPoint;
+import domain.card.Card;
+import domain.card.CardName;
 
 public class CardPointCalculator {
     public static CardPoint calculate(Card card) {
