+++ src/main/java/domain/blackjack/GameResult.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.blackjack;
 
 public enum GameResult {
     WIN, LOSE, TIE;
