+++ src/main/java/domain/blackjack/SummationCardPoint.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.blackjack;
 
 import java.util.List;
