+++ src/main/java/domain/card/CardDrawStrategy.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.card;
 
 import java.util.List;
