+++ src/test/java/domain/blackjack/SummationCardPointTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.blackjack;
 
 import java.util.List;
 import org.assertj.core.api.Assertions;
