+++ src/test/java/domain/blackjack/TestDealerCardDrawStrategy.java
@@ -1,5 +1,9 @@
-package domain;
+package domain.blackjack;
 
+import domain.blackjack.Gamer;
+import domain.blackjack.SummationCardPoint;
+import domain.card.Card;
+import domain.card.CardDrawStrategy;
 import java.util.List;
 
 public class TestDealerCardDrawStrategy implements CardDrawStrategy {
